Fixes restart issue, changes move logic
Reviewer: The move counter logic requires one modification: the number of moves should be incremented when 2 cards are clicked, not one. LM: Fixed this so that move increments when each match is checked. 

Reviewer: The restart functionality requires just one modification: when I click on a card and restart the game, then the card clicked in the new game tries to match itself with the card in the last game. You can fix this by clearing your list of open cards every time the game is started/restarted. LM: Fixed so that .active class removed from all cards upon reset.